requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
aiohttp
requests